factor out the editorCapturePlugin
